refactor: 메소드를 만들어 조건식에 넣어줌으로서 메소드의 이름을 통해 의도 들어내기

+++ src/main/java/lotto/model/domain/WinningStatistics.java
@@ -37,7 +37,7 @@ public class WinningStatistics {
     }
 
     private double sumOfReward() {
-        return (float) rankAndRankCount.keySet()
+        return (double) rankAndRankCount.keySet()
                 .stream()
                 .map(this::sumOfOneRankReward)
                 .reduce(Integer::sum)
